CAMBIOS

+++ src/php/models/mTematicas.php
@@ -52,68 +52,111 @@
             return $datos;
             }
         }
-
+        function mObtenerTematicayPersonajes($id_tematica){
+            $sql_tematica = "SELECT id_tematica, nombre as nombre_tematica FROM Tematica";
+            $resultado_tematica = $this->conexion->query($sql_tematica);
+        
+            $sql_personajes = "SELECT personaje.id_personaje, personaje.nombre as nombre_personaje, descripcion, imagen, ambito.id_ambito, ambito.nombre as nombre_ambito FROM Personaje AS personaje
+            INNER JOIN Tematica_Ambito_Personaje as tem_amb_pers ON personaje.id_personaje = tem_amb_pers.id_personaje 
+            INNER JOIN Ambito as ambito ON ambito.id_ambito=tem_amb_pers.id_ambito
+            WHERE tem_amb_pers.id_tematica = $id_tematica";
+            $resultado_personajes = $this->conexion->query($sql_personajes);
+        
+            if ($resultado_tematica->num_rows > 0) {
+                $datos_tematica = [];
+                
+                // Agregar información de la temática al array
+                while($fila_tematica = $resultado_tematica->fetch_assoc()){
+                    $datos_tematica[] = $fila_tematica;
+                }
+                
+        
+                // Agregar información de los personajes al array
+                $datos_personajes = [];
+                while ($fila_personajes = $resultado_personajes->fetch_assoc()) {
+                    $datos_personajes[] = $fila_personajes;
+                }
+                $datos_tematica['personajes'] = $datos_personajes;
+        
+                return $datos_tematica;
+            } else {
+                throw new Exception("Error al obtener datos de la temática y personajes: " . $this->conexion->error);
+            }
+        }
+        
+        function mModificarTematicayPersonajes($id_tematica, $nombre_tematica, $personajes){
+            $sql_modificar_tematica = "UPDATE Tematica SET nombre = '$nombre_tematica' WHERE id_tematica = '$id_tematica'";
+            $this->conexion->query($sql_modificar_tematica);
+        
+            foreach ($personajes as $i => $personaje) {
+                if (isset($personaje["imagen_$i"]) && file_exists($personaje["imagen_$i"]["tmp_name"])) {
+                    $imagen = $personaje["imagen_$i"];
+                    $ext = pathinfo($imagen["name"], PATHINFO_EXTENSION);
+                    $nombreimagen = uniqid() . "." . $ext;
+                    $carpeta_final = __DIR__ . "/../../img";
+                    $ruta_inicial = $imagen["tmp_name"];
+                    $ruta_final = $carpeta_final . DIRECTORY_SEPARATOR . $nombreimagen;
+                    move_uploaded_file($ruta_inicial, $ruta_final);
+                } else {
+                    $nombreimagen = null;
+                }
+                $id_personaje = $personaje['id_personaje'];
+                $nombre_personaje = $personaje['nombre_personaje'];
+                $descripcion = $personaje['descripcion'];
+        
+                $sql_modificar_personaje = "UPDATE Personaje SET nombre = '$nombre_personaje', descripcion='$descripcion', imagen = '$nombreimagen' WHERE id_personaje = '$id_personaje'";
+                
+                $this->conexion->query($sql_modificar_personaje);
+                
+            }
+        }
+        
         function mAltaTematicas($nombretematica, $personajes){
-            $nombretematica = ($nombretematica === '') ? "NULL" : "'" . $this->conexion->real_escape_string($nombretematica) . "'";
+            $nombretematica = empty($nombretematica) ? "NULL" : "'" . $this->conexion->real_escape_string($nombretematica) . "'";
             
             $this->conexion->begin_transaction();
-    
+        
             try {
                 $sql_insertar_tematica = "INSERT INTO Tematica(nombre) VALUES ($nombretematica)";
                 $this->conexion->query($sql_insertar_tematica);        
                 $id_tematica = $this->conexion->insert_id;
+        
                 // Crear personajes asociados a la temática
                 foreach ($personajes as $i => $personaje) {
-                    $ambito = $this->conexion->real_escape_string($personaje["ambito_$i"]);
-                    $id_ambito = (int)$ambito;
+                    $id_ambito = (int)$personaje["ambito_$i"];
                     $nombrepersonaje = $this->conexion->real_escape_string($personaje["nombre_personaje_$i"]);
                     $descripcion = $this->conexion->real_escape_string($personaje["descripcion_$i"]);
                     
                     if (isset($personaje["imagen_$i"]) && file_exists($personaje["imagen_$i"]["tmp_name"])) {
                         $imagen = $personaje["imagen_$i"];
                         $ext = pathinfo($imagen["name"], PATHINFO_EXTENSION);
-                        print_r($nombreimagen = uniqid() . "." . $ext);
+                        $nombreimagen = uniqid() . "." . $ext;
                         $carpeta_final = __DIR__ . "/../../img";
                         $ruta_inicial = $imagen["tmp_name"];
                         $ruta_final = $carpeta_final . DIRECTORY_SEPARATOR . $nombreimagen;
+                        
+                        move_uploaded_file($ruta_inicial, $ruta_final);
                     } else {
                         $nombreimagen = null;
                     }
-                
-                    // La siguiente línea incluye "NULL" para el campo imagen. Asegúrate de ajustar esto según tu estructura real de base de datos.
+                    
                     $sql_insertar_personaje = "INSERT INTO Personaje (nombre, descripcion, imagen) VALUES ('$nombrepersonaje', '$descripcion', '$nombreimagen')";
                     $this->conexion->query($sql_insertar_personaje);
                     $id_personaje = $this->conexion->insert_id;
-                    move_uploaded_file($ruta_inicial, $ruta_final);
+        
                     $sql_insertar_todo = "INSERT INTO Tematica_Ambito_Personaje(id_tematica, id_ambito, id_personaje) VALUES ('$id_tematica', '$id_ambito', '$id_personaje')";
                     $this->conexion->query($sql_insertar_todo);
-                }               
+                }
+        
                 $this->conexion->commit();
                 
             } catch (Exception $e) {
                 $this->conexion->rollback();
                 throw $e;
             }
-            
         }
-        /*function mModificarTematicas($id_tematica, $nombretematica, $personajes){
-            $nombretematica = ($nombretematica === '') ? "NULL" : "'" . $this->conexion->real_escape_string($nombretematica) . "'";
-            
-            $this->conexion->begin_transaction();
-            $sql = "UPDATE nombre FROM Tematica WHERE id_tematica = '$id_tematica'";
-            $this->conexion->query($sql);
+        
 
-            foreach ($personajes as $i => $personaje) {
-                $ambito = $this->conexion->real_escape_string($personaje["ambito_$i"]);
-                $id_ambito = (int)$ambito;
-                $nombrepersonaje = $this->conexion->real_escape_string($personaje["nombre_personaje_$i"]);
-                $descripcion = $this->conexion->real_escape_string($personaje["descripcion_$i"]);
-
-
-            //Nos redireccionamos de nuevo al listar tematicas, por si queremos borra una nueva tematica
-            header("Location: index.php?c=cTematicas&m=listarTematicas");
-            exit();
-        }*/
         function mBorrarTematicas($id_tematica){
             $sql = "SELECT personaje.id_personaje, personaje.imagen as imagen
             FROM Personaje AS personaje
